renamed classes and methods. Also merged two enrichers/rewriters to one, it appeared as helpful.
