+++ monitor/services/src/main/java/org/datacleaner/monitor/server/components/InputRewriter.java
@@ -22,6 +22,6 @@ package org.datacleaner.monitor.server.components;
 import org.datacleaner.descriptors.TransformerDescriptor;
 import org.datacleaner.restclient.ProcessStatelessInput;
 
-public interface InputEnricher {
-    boolean enrichStatelessInputForTransformer(TransformerDescriptor transformer, ProcessStatelessInput input);
+public interface InputRewriter {
+    boolean rewriteInput(TransformerDescriptor transformer, ProcessStatelessInput input);
 }
